Migrate to Netty 4.1 (#4397)
* Migrate to Netty 4.1

* Revert changes to vendor-module-ids.json and yarn.lock

Ceterum censeo JavaScript esse delendam.

* Add support for native SSL (OpenSSL) bindings

* Wait for transport being started completely in UdpTransportTest

* Refactor getBaseChannelHandlers/getFinalChannelHandlers

* Upgrade to Netty 4.1.18.Final

* Fix throughput and connection counters for Netty-based transports

* Use read bytes for CodecAggregator.Result in GelfChunkAggregator

* Refactor NettyTransport#getLocalAddress()

* Upgrade to Netty 4.1.19.Final

* Allow configuring the number of worker threads per Netty transport

* Remove support for Netty OIO (old I/O) transports

* Remove outdated TODOs

* Use locale metric registry for transport-specific metrics

+++ graylog2-server/src/main/java/org/graylog2/plugin/inputs/MessageInput.java
@@ -61,7 +61,7 @@ public abstract class MessageInput implements Stoppable {
     public static final String FIELD_CONTENT_PACK = "content_pack";
 
     @SuppressWarnings("StaticNonFinalField")
-    private static long defaultRecvBufferSize = 1024 * 1024;
+    private static int defaultRecvBufferSize = 1024 * 1024;
 
     private final MetricRegistry metricRegistry;
     private final Transport transport;
@@ -121,11 +121,11 @@ public abstract class MessageInput implements Stoppable {
         emptyMessages = localRegistry.counter("emptyMessages");
     }
 
-    public static long getDefaultRecvBufferSize() {
+    public static int getDefaultRecvBufferSize() {
         return defaultRecvBufferSize;
     }
 
-    public static void setDefaultRecvBufferSize(long size) {
+    public static void setDefaultRecvBufferSize(int size) {
         defaultRecvBufferSize = size;
     }
